requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
deepdiff==5.7.0
execnet==1.9.0
filelock==3.4.2
h11==0.13.0
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
jsonschema==4.4.0
login==0.0.6
ordered-set==4.0.2
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==6.2.5
pytest-assume==2.4.3
pytest-forked==1.4.0
pytest-xdist==2.5.0
PyYAML==6.0
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
zipp==3.7.0